[Refactor] [FE-44] 폴더 구조 변경
