+++ src/screens/bottomNav/MyPage/Setting/InquiryDetails/InquiryDetailsScreen.tsx
